+++ WONNIT/Features/SpaceActions/CreateSpace/FormSteps/CreateSpaceFormStep.swift
@@ -8,15 +8,13 @@
 import Foundation
 import SwiftUI
 
-enum CreateSpaceFormStep: CaseIterable, Identifiable {
+enum CreateSpaceFormStep: FormStep {
     case addressAndName
     case pictures
     case categoryAndTags
     case operation
     case scanner
     case miscellaneous
-    
-    var id: String { "\(self)" }
     
     var sectionTitle: String {
         switch self {
@@ -44,15 +42,113 @@ enum CreateSpaceFormStep: CaseIterable, Identifiable {
         }
     }
     
-    var next: CreateSpaceFormStep? {
-        guard let idx = Self.allCases.firstIndex(of: self),
-              idx + 1 < Self.allCases.count else { return nil }
-        return Self.allCases[idx + 1]
-    }
-    
-    var previous: CreateSpaceFormStep? {
-        guard let idx = Self.allCases.firstIndex(of: self),
-              idx > 0 else { return nil }
-        return Self.allCases[idx - 1]
+    var components: [FormComponent] {
+        switch self {
+        case .addressAndName:
+            return [
+                .textField(config: .init(
+                    id: "address1",
+                    title: "주소",
+                    placeholder: "주소를 입력해주세요"
+                )),
+                .textField(config: .init(
+                    id: "address2",
+                    title: "상세주소",
+                    placeholder: "상세주소를 입력해주세요"
+                )),
+                .textField(config: .init(
+                    id: "name",
+                    title: "공간이름",
+                    placeholder: "공간 이름을 입력해주세요 / 예) 노원 힐링스페이스"
+                )),
+                .doubleField(
+                    config: .init(
+                        id: "area",
+                        title: "공간 크기",
+                        placeholder: "공간 크기를 입력해주세요 / 예) 12.25",
+                        suffix: "m²",
+                        keyboardType: .decimalPad
+                    )
+                )
+            ]
+
+        case .pictures:
+            return [
+                .imageUploaderWithML(
+                    config: .init(
+                        id: "mainImage",
+                        title: "대표사진 등록",
+                        spaceCategoryFormComponentKey: "category",
+                        spaceTagFormComponentKey: "tags"
+                    ),
+                    variant: .singleLarge
+                ),
+                .imageUploaderWithML(
+                    config: .init(
+                        id: "subImages",
+                        title: "추가사진 등록",
+                        spaceTagFormComponentKey: "tags"
+                    ),
+                    variant: .multipleSmall(limit: 4)
+                )
+            ]
+            
+        case .categoryAndTags:
+            return [
+                .select(
+                    config: .init(
+                        id: "category",
+                        title: "공간 카테고리",
+                        placeholder: "카테고리를 선택하세요",
+                        isAIFeatured: true
+                    ),
+                    options: SpaceCategory.allCases.map(\.label)
+                ),
+                .tagSelector(config: .init(
+                    id: "tags",
+                    title: "구비물품",
+//                    placeholder: "구비 물품을 작성하세요",
+                    isAIFeatured: true,
+                    spaceCategoryFormComponentKey: "category"
+                ))
+            ]
+
+        case .operation:
+            return [
+                .dayPicker(config: .init(
+                    id: "openDay",
+                    title: "운영요일"
+                )),
+                .timeRangePicker(config: .init(
+                    id: "openTime",
+                    title: "운영시간"
+                )),
+                .pricingField(config: .init(
+                    id: "pricing",
+                    title: "금액정보"
+                ))
+            ]
+
+        case .scanner:
+            return [
+                .roomScanner(config: .init(
+                    id: "roomScanner",
+                ))
+            ]
+
+        case .miscellaneous:
+            return [
+                .phoneNumberField(config: .init(
+                    id: "contact",
+                    title: "담당자 연락처",
+                    placeholder: "연락처를 입력해주세요",
+                )),
+                .textField(config: .init(
+                    id: "cautions",
+                    title: "기타 주의사항(선택)",
+                    placeholder: "주의사항이 있다면 입력해주세요"
+                ))
+            ]
+        }
     }
 }
